fix: cpu frequency cannot show on ksystemstats 5.27
Tested on ksystemstats 5.27.11(kubuntu24.04), there are "cpu/all/averageFrequency" which didn't
exists in previous version. So the code being deleted in this commit is executed, which disable the
CpuFrequency.qml component.

In fact, in this version, "cpu/all/averageFrequency" is never directly used even if it exists. So
just delete it.